Struct umya_spreadsheet::structs::Worksheet

pub struct Worksheet { /* private fields */ }

A Worksheet Object.

Implementations

impl Worksheet

pub fn get_value(&self, coordinate: &str) -> String

Get value.

Arguments

coordinate - Specify the coordinates. ex) “A1”

Return value

String - Value of the specified cell.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let value = worksheet.get_value("A1");

pub fn get_value_by_column_and_row(&self, col: &u32, row: &u32) -> String

Get value by specifying the column number and row number.

Arguments

col - Specify the column number. (first column number is 1)

row - Specify the row number. (first row number is 1)

Return value

String - Value of the specified cell.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let value = worksheet.get_value_by_column_and_row(&1, &1);

pub fn get_formatted_value(&self, coordinate: &str) -> String

Get formatted value.

Arguments

coordinate - Specify the coordinates. ex) “A1”

Return value

String - Formatted value of the specified cell.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let value = worksheet.get_formatted_value("A1");

pub fn get_formatted_value_by_column_and_row(
&self,
col: &u32,
row: &u32
) -> String

Get formatted value by specifying the column number and row number.

Arguments

col - Specify the column number. (first column number is 1)

row - Specify the row number. (first row number is 1)

Return value

String - Formatted value of the specified cell.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let value = worksheet.get_formatted_value_by_column_and_row(&1, &1);

pub fn get_cell_collection(&self) -> Vec<&Cell>

Get Cell List.

pub fn get_cell_collection_mut(&mut self) -> Vec<&mut Cell>

Get Cell List in mutable.

pub fn get_collection_to_hashmap(&self) -> &HashMap<(u32, u32), Cell>

pub fn get_collection_to_hashmap_mut(
&mut self
) -> &mut HashMap<(u32, u32), Cell>

pub fn get_cell(&self, coordinate: &str) -> Option<&Cell>

Get cell.

Arguments

coordinate - Specify the coordinates. ex) “A1”

Return value

Option - Cell in the Some.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let cell = worksheet.get_cell("A1");

pub fn get_cell_by_column_and_row(&self, col: &u32, row: &u32) -> Option<&Cell>

Gets the cell by specifying the column number and row number.

Arguments

col - Specify the column number. (first column number is 1)

row - Specify the row number. (first row number is 1)

Return value

Option - Cell in the Some.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let cell = worksheet.get_cell_by_column_and_row(&1, &1);  // get cell from A1.

pub fn get_cell_mut(&mut self, coordinate: &str) -> &mut Cell

Get cell with mutable.

Arguments

coordinate - Specify the coordinates. ex) “A1”

Return value

&mut Cell - Cell with mutable.

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
let cell = worksheet.get_cell_mut("A1");

pub fn get_cell_by_column_and_row_mut(
&mut self,
col: &u32,
row: &u32
) -> &mut Cell

Gets the cell with mutable by specifying the column number and row number.

Arguments

col - Specify the column number. (first column number is 1)

row - Specify the row number. (first row number is 1)

Return value

*&mut Cell - Cell with mutable.

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
let cell = worksheet.get_cell_by_column_and_row_mut(&1, &1);  // get cell from A1.

pub fn get_collection_by_column(&self, column_num: &u32) -> Vec<&Cell>

pub fn get_collection_by_row(&self, row_num: &u32) -> Vec<&Cell>

pub fn get_collection_by_column_to_hashmap(
&self,
column_num: &u32
) -> HashMap<u32, &Cell>

pub fn get_collection_by_row_to_hashmap(
&self,
row_num: &u32
) -> HashMap<u32, &Cell>

pub fn set_cell(&mut self, cell: Cell) -> &mut Self

Set Cell

Arguments

cell - Cell

pub fn remove_cell_by_column_and_row_mut(
&mut self,
col: &u32,
row: &u32
) -> bool

pub fn get_cell_value(&self, coordinate: &str) -> &CellValue

Get cell value.

Arguments

coordinate - Specify the coordinates. ex) “A1”

Return value

&CellValue - CellValue.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let cell_value = worksheet.get_cell_value("A1");

pub fn get_cell_value_by_column_and_row(
&self,
col: &u32,
row: &u32
) -> &CellValue

Gets the cell value by specifying the column number and row number.

Arguments

col - Specify the column number. (first column number is 1)

row - Specify the row number. (first row number is 1)

Return value

&CellValue - CellValue.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let cell_value = worksheet.get_style_by_column_and_row(&1, &1);  // get cell from A1.

pub fn get_cell_value_mut(&mut self, coordinate: &str) -> &mut CellValue

Get cell value with mutable.

Arguments

coordinate - Specify the coordinates. ex) “A1”

Return value

&mut CellValue - CellValue with mutable.

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
let cell_value = worksheet.get_cell_value_mut("A1");

pub fn get_cell_value_by_column_and_row_mut(
&mut self,
col: &u32,
row: &u32
) -> &mut CellValue

Gets the cell value with mutable by specifying the column number and row number.

Arguments

col - Specify the column number. (first column number is 1)

row - Specify the row number. (first row number is 1)

Return value

*&mut CellValue - CellValue with mutable.

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
let cell_value = worksheet.get_cell_value_by_column_and_row_mut(&1, &1);  // get cell_value from A1.

pub fn get_cell_value_by_range(&self, range: &str) -> Vec<&CellValue>

Gets the cell value by specifying an range.

Arguments

range - range. ex) “A1:C5”

Return value

*Vec<&CellValue> - CellValue List.

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
let mut cell_value_List = worksheet.get_cell_value_by_range("A1:C5");

pub fn get_style(&self, coordinate: &str) -> &Style

Get style.

Arguments

coordinate - Specify the coordinates. ex) “A1”

Return value

&Style - Style.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let style = worksheet.get_style("A1");

pub fn get_style_by_column_and_row(&self, col: &u32, row: &u32) -> &Style

Gets the style by specifying the column number and row number.

Arguments

col - Specify the column number. (first column number is 1)

row - Specify the row number. (first row number is 1)

Return value

&Style - Style.

Examples

let book = umya_spreadsheet::new_file();
let worksheet = book.get_sheet(&0).unwrap();
let style = worksheet.get_style_by_column_and_row(&1, &1);  // get cell from A1.

pub fn get_style_mut(&mut self, coordinate: &str) -> &mut Style

Get style with mutable.

Arguments

coordinate - Specify the coordinates. ex) “A1”

Return value

&mut Style - Style with mutable.

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
let style = worksheet.get_style_mut("A1");

pub fn get_style_by_column_and_row_mut(
&mut self,
col: &u32,
row: &u32
) -> &mut Style

Gets the style with mutable by specifying the column number and row number.

Arguments

col - Specify the column number. (first column number is 1)

row - Specify the row number. (first row number is 1)

Return value

*&mut Style - Style with mutable.

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
let style = worksheet.get_style_by_column_and_row_mut(&1, &1);  // get style from A1.

pub fn set_style(&mut self, coordinate: &str, style: Style) -> &mut Self

pub fn set_style_by_column_and_row(
&mut self,
col: &u32,
row: &u32,
style: Style
) -> &mut Self

Set the style by specifying the column number and row number.

Arguments

col - Specify the column number. (first column number is 1)

row - Specify the row number. (first row number is 1)

style - Style.

Return value

*&mut Self - self.

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
let mut style = umya_spreadsheet::Style::default();
style.get_borders_mut().get_bottom_mut().set_border_style(umya_spreadsheet::Border::BORDER_MEDIUM);
let style = worksheet.set_style_by_column_and_row(&1, &1, style);  // set style to A1.

pub fn set_style_by_range(&mut self, range: &str, style: Style) -> &mut Self

Set style by range.

Arguments

range - Specify the range. ex) “A1:B2”

style - Style

Return value

&mut Self - Self.

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
let mut style = umya_spreadsheet::Style::default();
style.get_borders_mut().get_bottom_mut().set_border_style(umya_spreadsheet::Border::BORDER_MEDIUM);
worksheet.set_style_by_range("A1:A3", style);

pub fn get_comments(&self) -> &Vec<Comment>

Get Comments

pub fn get_comments_mut(&mut self) -> &mut Vec<Comment>

Get Comments in mutable.

pub fn get_comments_to_hashmap(&self) -> HashMap<String, &Comment>

Get Comments convert to hashmap.

pub fn set_comments(&mut self, value: Vec<Comment>)

Set Comments.

Arguments

value - Comment List (Vec)

pub fn add_comments(&mut self, value: Comment)

Add Comments.

Arguments

value - Comment

pub fn has_comments(&self) -> bool

Has Comments.

pub fn get_conditional_styles_collection(&self) -> &Vec<ConditionalSet>

Get ConditionalSet list.

pub fn set_conditional_styles_collection(&mut self, value: Vec<ConditionalSet>)

Set ConditionalSet.

Arguments

value - ConditionalSet List (Vec)

pub fn add_conditional_styles_collection(&mut self, value: ConditionalSet)

Add ConditionalSet.

Arguments

value - ConditionalSet

pub fn get_merge_cells(&self) -> &Vec<Range>

pub fn get_merge_cells_mut(&mut self) -> &mut Vec<Range>

pub fn add_merge_cells<S: Into<String>>(&mut self, range: S) -> &mut Self

Arguments

range - Range. ex) “A1:C5”

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
worksheet.add_merge_cells("A1:C5");

pub fn get_auto_filter(&self) -> &Option<AutoFilter>

pub fn get_auto_filter_mut(&mut self) -> &mut Option<AutoFilter>

pub fn set_auto_filter<S: Into<String>>(&mut self, range: S)

Arguments

range - Range. ex) “A2:K2”

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
worksheet.set_auto_filter("A2:K2");

pub fn remove_auto_filter(&mut self)

pub fn get_column_dimensions(&self) -> &Vec<Column>

Get Column Dimension List.

pub fn get_column_dimensions_mut(&mut self) -> &mut Vec<Column>

Get Column Dimension List in mutable.

pub fn calculation_auto_width(&mut self) -> &mut Self

Calculation Auto Width.

pub fn get_column_dimension(&self, column: &str) -> Option<&Column>

Get Column Dimension.

Arguments

column - Column Char. ex) “A”

pub fn get_column_dimension_mut(&mut self, column: &str) -> &mut Column

Get Column Dimension in mutable.

Arguments

column - Column Char. ex) “A”

pub fn get_column_dimension_by_number(&self, col: &u32) -> Option<&Column>

Get Column Dimension.

Arguments

col - Column Number.

pub fn get_column_dimension_by_number_mut(&mut self, col: &u32) -> &mut Column

Get Column Dimension in mutable.

Arguments

col - Column Number.

pub fn has_sheet_data(&self) -> bool

pub fn get_row_dimensions(&self) -> Vec<&Row>

Get Row Dimension List.

pub fn get_row_dimensions_mut(&mut self) -> Vec<&mut Row>

Get Row Dimension List in mutable.

pub fn get_row_dimensions_to_hashmap(&self) -> &HashMap<u32, Row>

Get Row Dimension convert Hashmap.

pub fn get_row_dimensions_to_hashmap_mut(&mut self) -> &mut HashMap<u32, Row>

pub fn get_row_dimension(&self, row: &u32) -> Option<&Row>

Get Row Dimension.

pub fn get_row_dimension_mut(&mut self, row: &u32) -> &mut Row

Get Row Dimension in mutable.

pub fn get_worksheet_drawing(&self) -> &WorksheetDrawing

Get WorksheetDrawing.

pub fn get_worksheet_drawing_mut(&mut self) -> &mut WorksheetDrawing

Get WorksheetDrawing in mutable.

pub fn set_worksheet_drawing(&mut self, value: WorksheetDrawing)

Set WorksheetDrawing.

Arguments

value - WorksheetDrawing

pub fn has_drawing_object(&self) -> bool

Has WorksheetDrawing.

pub fn insert_new_row(&mut self, row_index: &u32, num_rows: &u32)

Insert new rows.

Arguments

row_index - Specify point of insert. ex) 1

num_rows - Specify number to insert. ex) 2

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
worksheet.insert_new_row(&2, &3);

pub fn insert_new_row_from_other_sheet(
&mut self,
sheet_name: &str,
row_index: &u32,
num_rows: &u32
)

Adjust for references to other sheets.

pub fn insert_new_column(&mut self, column: &str, num_columns: &u32)

Insert new columns.

Arguments

column - Specify point of insert. ex) “B”

num_columns - Specify number to insert. ex) 3

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
worksheet.insert_new_column("B", &3);

pub fn insert_new_column_from_other_sheet(
&mut self,
sheet_name: &str,
column: &str,
num_columns: &u32
)

Adjust for references to other sheets.

pub fn insert_new_column_by_index(
&mut self,
column_index: &u32,
num_columns: &u32
)

Insert new columns.

Arguments

column_index - Specify point of insert. ex) 2

num_columns - Specify number to insert. ex) 3

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
worksheet.insert_new_column_by_index(&2, &3);

pub fn insert_new_column_by_index_from_other_sheet(
&mut self,
sheet_name: &str,
column_index: &u32,
num_columns: &u32
)

Adjust for references to other sheets.

pub fn remove_row(&mut self, row_index: &u32, num_rows: &u32)

Remove rows.

Arguments

row_index - Specify point of remove. ex) 1

num_rows - Specify number to remove. ex) 2

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
worksheet.remove_row(&2, &3);

pub fn remove_row_from_other_sheet(
&mut self,
sheet_name: &str,
row_index: &u32,
num_rows: &u32
)

Adjust for references to other sheets.

pub fn remove_column(&mut self, column: &str, num_columns: &u32)

Remove columns.

Arguments

sheet_name - Specify the sheet name. ex) “Sheet1”

column - Specify point of remove. ex) “B”

num_columns - Specify number to remove. ex) 3

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
worksheet.remove_column("B", &3);

pub fn remove_column_from_other_sheet(
&mut self,
sheet_name: &str,
column: &str,
num_columns: &u32
)

Adjust for references to other sheets.

pub fn remove_column_by_index(&mut self, column_index: &u32, num_columns: &u32)

Remove columns.

Arguments

column_index - Specify point of remove. ex) 2

num_columns - Specify number to remove. ex) 3

Examples

let mut book = umya_spreadsheet::new_file();
let mut worksheet = book.get_sheet_mut(&0).unwrap();
worksheet.remove_column_by_index(&2, &3);

pub fn remove_column_by_index_from_other_sheet(
&mut self,
sheet_name: &str,
column_index: &u32,
num_columns: &u32
)

Adjust for references to other sheets.

pub fn get_code_name(&self) -> &Option<String>

Get Code Name.

pub fn set_code_name<S: Into<String>>(&mut self, value: S)

Set Code Name.

Arguments

value - Code Name

pub fn get_header_footer(&self) -> &HeaderFooter

Get Header Footer.

pub fn get_header_footer_mut(&mut self) -> &mut HeaderFooter

Get Header Footer in mutable.

pub fn set_header_footer(&mut self, value: HeaderFooter) -> &mut Self

Set Header Footer.

Arguments

value - Header Footer

pub fn get_active_cell(&self) -> &str

Get Active Cell.

pub fn set_active_cell<S: Into<String>>(&mut self, cell: S)

Set Active Cell.

Arguments

cell - Cell ex) “A1”

pub fn get_sheet_id(&self) -> &String

Get Sheet Id.

pub fn has_code_name(&self) -> bool

Has Code Name.

pub fn get_tab_color(&self) -> &Option<Color>

Get Tab Color.

pub fn get_tab_color_mut(&mut self) -> &mut Color

Get Tab Color in mutable.

pub fn set_tab_color(&mut self, value: Color) -> &mut Self

Set Tab Color.

Arguments

value - Color

pub fn remove_tab_color(&mut self) -> &mut Self

Remove Tab Color.

pub fn calculate_worksheet_dimension(&self) -> String

Calculate Worksheet Dimension.

pub fn get_highest_column_and_row(&self) -> (u32, u32)

Return value

*(u32, u32) - (column, row)

pub fn get_highest_column(&self) -> u32

pub fn get_highest_row(&self) -> u32

pub fn get_title(&self) -> &str

👎 Deprecated since 1.0.0:
please use get_name instead

Get Title.

pub fn get_name(&self) -> &str

Get SheetName.

pub fn set_title<S: Into<String>>(&mut self, sheet_title: S) -> &mut Self

👎 Deprecated since 1.0.0:
please use set_name instead

Set Title.

Arguments

sheet_title - Sheet Title. [Caution] no duplicate other worksheet.

pub fn set_name<S: Into<String>>(&mut self, sheet_name: S) -> &mut Self

Set SheetName.

Arguments

sheet_name - Sheet Name. [Caution] no duplicate other worksheet.

pub fn get_sheet_state(&self) -> &str

pub fn set_sheet_state(&mut self, value: String) -> &mut Self

Set Sheet State.

Arguments

value - Sheet State.

pub fn get_page_setup(&self) -> &PageSetup

pub fn get_page_setup_mut(&mut self) -> &mut PageSetup

pub fn set_page_setup(&mut self, value: PageSetup) -> &mut Self

Set Page Setup.

Arguments

value - PageSetup.

pub fn get_page_margins(&self) -> &PageMargins

pub fn get_page_margins_mut(&mut self) -> &mut PageMargins

pub fn set_page_margins(&mut self, value: PageMargins) -> &mut Self

Set Page Margins.

Arguments

value - PageMargins.

pub fn get_sheets_views(&self) -> &SheetViews

pub fn get_sheet_views_mut(&mut self) -> &mut SheetViews

pub fn set_sheets_views(&mut self, value: SheetViews) -> &mut Self

Set SheetViews.

Arguments

value - SheetViews.

pub fn get_ole_objects(&self) -> &OleObjects

pub fn get_ole_objects_mut(&mut self) -> &mut OleObjects

pub fn set_ole_objects(&mut self, value: OleObjects) -> &mut Self

Set Ole Objects.

Arguments

value - OleObjects.

pub fn get_defined_names(&self) -> &Vec<DefinedName>

Get Defined Name (Vec).

pub fn get_defined_names_mut(&mut self) -> &mut Vec<DefinedName>

Get Defined Name (Vec) in mutable.

pub fn set_defined_names(&mut self, value: Vec<DefinedName>)

Set Defined Name (Vec).

Arguments

value - Vec.

pub fn add_defined_names(&mut self, value: DefinedName)

Add Defined Name.

Arguments

value - DefinedName.

pub fn add_defined_name<S: Into<String>>(
&mut self,
name: S,
address: S
) -> Result<(), &str>

Add Defined Name.

Arguments

name - Name. ex) “DefinedName01”

address - Address. ex) “A1:A2”

pub fn get_print_options(&self) -> &PrintOptions

Get Print Options.

pub fn get_print_options_mut(&mut self) -> &mut PrintOptions

Get Print Options in mutable.

pub fn set_print_options(&mut self, value: PrintOptions) -> &mut Self

Set Print Options.

Arguments

value - PrintOptions.

pub fn get_column_breaks(&self) -> &ColumnBreaks

Get Column Breaks.

pub fn get_column_breaks_mut(&mut self) -> &mut ColumnBreaks

Get Column Breaks in mutable.

pub fn set_column_breaks(&mut self, value: ColumnBreaks) -> &mut Self

Set Column Breaks.

Arguments

value - ColumnBreaks.

pub fn get_row_breaks(&self) -> &RowBreaks

Get Row Breaks.

pub fn get_row_breaks_mut(&mut self) -> &mut RowBreaks

Get Row Breaks in mutable.

pub fn set_row_breaks(&mut self, value: RowBreaks) -> &mut Self

Set Row Breaks.

Arguments

value - RowBreaks.

pub fn get_image_collection(&self) -> &Vec<Image>

Outputs all images contained in the worksheet.

Return value

&Vec<Image> - Image Object List.

pub fn get_image_collection_mut(&mut self) -> &mut Vec<Image>

Outputs all images contained in the worksheet.

Return value

&mut Vec<Image> - Image Object List.

pub fn add_image(&mut self, value: Image) -> &mut Self

pub fn get_image(&self, coordinate: &str) -> Option<&Image>

pub fn get_image_by_column_and_row(
&self,
col: &u32,
row: &u32
) -> Option<&Image>

pub fn get_image_mut(&mut self, coordinate: &str) -> Option<&mut Image>

pub fn get_image_by_column_and_row_mut(
&mut self,
col: &u32,
row: &u32
) -> Option<&mut Image>

pub fn get_images(&self, coordinate: &str) -> Vec<&Image>

pub fn get_images_by_column_and_row(&self, col: &u32, row: &u32) -> Vec<&Image>

pub fn get_images_mut(&mut self, coordinate: &str) -> Vec<&mut Image>

pub fn get_images_by_column_and_row_mut(
&mut self,
col: &u32,
row: &u32
) -> Vec<&mut Image>

pub fn get_chart_collection(&self) -> &Vec<Chart>

Outputs all Charts contained in the worksheet.

Return value

&Vec<Chart> - Chart Object List.

pub fn get_chart_collection_mut(&mut self) -> &mut Vec<Chart>

Outputs all Charts contained in the worksheet.

Return value

&mut Vec<Chart> - Chart Object List.

pub fn add_chart(&mut self, value: Chart) -> &mut Self

pub fn get_chart(&self, coordinate: &str) -> Option<&Chart>

pub fn get_chart_by_column_and_row(
&self,
col: &u32,
row: &u32
) -> Option<&Chart>

pub fn get_chart_mut(&mut self, coordinate: &str) -> Option<&mut Chart>

pub fn get_chart_by_column_and_row_mut(
&mut self,
col: &u32,
row: &u32
) -> Option<&mut Chart>

pub fn get_charts(&self, coordinate: &str) -> Vec<&Chart>

pub fn get_charts_by_column_and_row(&self, col: &u32, row: &u32) -> Vec<&Chart>

pub fn get_charts_mut(&mut self, coordinate: &str) -> Vec<&mut Chart>

pub fn get_charts_by_column_and_row_mut(
&mut self,
col: &u32,
row: &u32
) -> Vec<&mut Chart>